requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
Flask==3.0.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
WTForms==3.1.2
Werkzeug==3.1.6

# ── Database ───────────────────────────────────────────────────────────────────
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.10
alembic==1.14.0

# ── Redis / Sessions ───────────────────────────────────────────────────────────
redis==5.2.1
Flask-Session==0.8.0
cachelib==0.13.0

# ── Scheduling ─────────────────────────────────────────────────────────────────
APScheduler==3.10.4

# ── HTTP / API ─────────────────────────────────────────────────────────────────
requests==2.32.3
urllib3==2.2.3

# ── Document parsing ───────────────────────────────────────────────────────────
pdfplumber==0.11.4
pandas==2.2.3
numpy==2.4.2
openpyxl==3.1.5
xlrd==2.0.1
pytesseract==0.3.13
Pillow==11.0.0

# ── NFL data (seed scripts) ────────────────────────────────────────────────────
nfl_data_py==0.3.3
pyarrow==23.0.1
fastparquet==2025.12.0

# ── Report export ──────────────────────────────────────────────────────────────
reportlab==4.2.5

# ── Logging ────────────────────────────────────────────────────────────────────
python-json-logger==2.0.7

# ── Production server ──────────────────────────────────────────────────────────
gunicorn==23.0.0

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
click==8.1.7
python-dateutil==2.9.0.post0
pytz==2025.2
